+++ test/70-auto.spec.js
@@ -4,7 +4,6 @@
 
 const { assert } = require('chai');
 const { v4: uuid } = require('uuid');
-const Promise = require('bluebird');
 const cts = require('./challtestsrv');
 const getCertIssuers = require('./get-cert-issuers');
 const spec = require('./spec');
@@ -267,7 +266,7 @@ describe('client.auto', () => {
             this.skip();
         }
 
-        await Promise.map(testIssuers, async (issuer) => {
+        await Promise.all(testIssuers.map(async (issuer) => {
             const [, csr] = await acme.forge.createCsr({
                 commonName: `${uuid()}.${domainName}`
             });
@@ -284,7 +283,7 @@ describe('client.auto', () => {
             const info = await acme.forge.readCertificateInfo(rootCert);
 
             assert.strictEqual(issuer, info.issuer.commonName);
-        });
+        }));
     });
 
     it('should get default chain with invalid preference [ACME_CAP_ALTERNATE_CERT_ROOTS]', async function() {