+++ test/get-cert-issuers.js
@@ -2,7 +2,6 @@
  * Get ACME certificate issuers
  */
 
-const Promise = require('bluebird');
 const acme = require('./../');
 const util = require('./../src/util');
 
@@ -17,11 +16,11 @@ async function getPebbleCertIssuers() {
     /* Get intermediate certificate and resolve alternates */
     const root = await acme.axios.get(`${pebbleManagementUrl}/intermediates/0`);
     const links = util.parseLinkHeader(root.headers.link || '');
-    const alternates = await Promise.map(links, async (link) => acme.axios.get(link));
+    const alternates = await Promise.all(links.map(async (link) => acme.axios.get(link)));
 
     /* Get certificate info */
     const certs = [root].concat(alternates).map((c) => c.data);
-    const info = await Promise.map(certs, acme.forge.readCertificateInfo);
+    const info = await Promise.all(certs.map(async (c) => acme.forge.readCertificateInfo(c)));
 
     /* Return issuers */
     return info.map((i) => i.issuer.commonName);