+++ CHANGELOG.md
@@ -11,7 +11,7 @@ A more detailed explanation can be found [at the Let's Encrypt forums](https://c
 
 * `added` New native crypto engine, ECC/ECDSA support
 * `changed` Replace `bluebird` dependency with native promise APIs
-* `breaking` Remove support for Node v10
+* `breaking` Remove support for Node v10, v12 and v14
 
 
 ## v4.2.5 (2022-03-21)